+++ TEST_HERE/backend/pjt/urls.py
@@ -1,19 +1,3 @@
-"""
-URL configuration for pjt project.
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/4.2/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.contrib import admin
 from django.urls import path, include
 from django.conf import settings
@@ -21,7 +5,7 @@ from django.conf.urls.static import static
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/articles/', include('articles.urls')),  # articles 앱 URL 연결
+    path('api/stocks/', include('stocks.urls')), # 이거 순서 조심해야 함
     path('accounts/', include('accounts.urls')),
     path('api/auth/', include('accounts.urls')),  # accounts의 사용자 정보 URL 포함
     path('', include('stocks.urls')),  # stocks 앱의 URL 포함
